feat: 구조 개선 및 ui 버그 수정
- 제목 누르면 refresh 하도록 변경
- message input box 개선
- 기본적인 채팅 ui 완성
