+++ client/src/containers/home/navigator.tsx
@@ -2,7 +2,7 @@ import Image from "next/image"
 
 function NavFooterTool({ iconSrc, text, link }: { iconSrc: string; text: string; link: string }) {
   return (
-    <a className="flex mt-7 ml-2.5" href={link}>
+    <a className="flex mt-7 ml-2.5 items-center" href={link}>
       <Image src={iconSrc} alt="" height="16" width="16" style={{ height: "16px" }} />
       <p className="ml-2 text-sm">{text}</p>
     </a>
