+++ client/src/styles/globals.css
@@ -24,14 +24,15 @@
         background-color: rgb(0,0,0,0.2);
         border-radius: 100px;
     }
-    .custom-scroll-bar-12px::-webkit-scrollbar-thumb:hover {
+    .custom-scroll-bar-10px::-webkit-scrollbar-thumb:hover {
         background-color: rgb(0,0,0,0.3);
     }
-    .custom-scroll-bar-12px::-webkit-scrollbar {
-        width: 12px;
-        height: 12px;
+    .custom-scroll-bar-10px::-webkit-scrollbar {
+        width: 10px;
+        height: 10px;
+
     }
-    .custom-scroll-bar-12px::-webkit-scrollbar-thumb {
+    .custom-scroll-bar-10px::-webkit-scrollbar-thumb {
         background-color: rgb(0,0,0,0.2);
         border-radius: 100px;
     }
